Minor but essential changes due to move from under ExtensionPointWizardPage to under GeneratedComponentWizardPage.
Added fields to represent wizard name and description (as is done in other subtypes of GeneratedComponentWizardPage).

Adapted constructor to better fit with new supertype constructor.

Expanded body of createAdditionalControls(..) to explicitly create the "class" field on the wizard page (which was previously addressed through the schema that is being dropped).

Other minor edits of code, format, and comments.